Address low and moderate security concerns
Use Red Hat UBI NodeJS 18 image as builder image in the Dockerfile.
This ensures that all images benefit from updates by Red Hat.

Add the USER statement to the Dockerfile to ensure a non root user
will run the application. Here 1001 is the default for UBI images.

Install NGINX on top of UBI 8 Minimal to reduce the image size and
possibly the attack surface. The size decreases from 452 MB to 216 MB.

Related to PLMPGM-829.

Signed-off-by: Fabien Dupont <fabiendupont@pm.me>