Add DCGM exporter metrics ConfigMap to Helm Chart
This change adds the DCGM exporter metrics ConfigMap to the Helm Chart
templates, in order to be deployed alongside the Console Plugin. The
metrics included are the required metrics for the NVIDIA GPU Admin
Dashboard to be fully functional.

The name of this ConfigMap should be included in the NVIDIA operator ClusterPolicy
CR, under `.spec.dcgmExporter.config.name`, for all the metrics to be
exported as expected.

Signed-off-by: Michail Resvanis <mresvani@redhat.com>