Auto-generate metrics config by merging with upstream
Users of this console plugin are required to replace their
current list of DCGM metrics with the list installed by the
plugin, because the plugin uses a few metrics that are not
included with the DCGM exporter by default.

In order to make this configuration switch less intrusive,
and to cover the most likely case when no custom DCGM
configuration is used, we want to include the metrics required
by this plugin in addition to, and not instead of, the default ones.